+++ test/task5/ContactRepositoryTest.java
@@ -1,7 +1,9 @@
-import com.clouway.adapter.ConnectionProvider;
-import com.clouway.adapter.PersistentContactRepository;
-import com.clouway.core.Contact;
-import com.clouway.datastore.DataStore;
+package task5;
+
+import com.clouway.task5.adapter.ConnectionProvider;
+import com.clouway.task5.adapter.PersistentContactRepository;
+import com.clouway.task5.adapter.core.Contact;
+import com.clouway.task5.adapter.datastore.DataStore;
 import org.junit.Before;
 import org.junit.Test;
